* master:
  Remove ge_scalarmult_vartime() which is not used any more
  edwards25519sha512batch_open(): check order before decoding instead of after
  Accept non-canonical PKs if ED25519_COMPAT is defined
  ed25519_open(): reject all small order public keys and non-canonical representations
  edwards25519sha512batch: reuse ge_scalarmult()
  Add constant-time edx recovery; use it in ed25519_scalarmult()
  Don't hardcode the number of rounds
  CMake find module fixes for static linking: (#621)
  blake2: use the same code on little & big endian for finalization
  Use uint instead of uint64_t for SHA* padding
  Use the output buffer as a temporary buffer to store clamped private scalars
  Return -1 is the scalar was zero
  Clear the high bit the same way everywhere
  Remove unused var
  Add a comment on scalarmult usage

+++ src/libsodium/crypto_hash/sha256/cp/hash_sha256_cp.c
@@ -153,10 +153,10 @@ static const uint8_t PAD[64] = { 0x80, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
 static void
 SHA256_Pad(crypto_hash_sha256_state *state, uint32_t tmp32[64 + 8])
 {
-    uint64_t r;
-    uint64_t i;
+    unsigned int r;
+    unsigned int i;
 
-    r = (state->count >> 3) & 0x3f;
+    r = (unsigned int) ((state->count >> 3) & 0x3f);
     if (r < 56) {
         for (i = 0; i < 56 - r; i++) {
             state->buf[r + i] = PAD[i];
